+++ CHANGELOG.md
@@ -7,11 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- add share dropdown menu and functional QR code generator to game landing invite modal (ARC-958)
+
 ## [1.29.32] - 2026-09-11
 
 ### Added
-- add public player profiles with game history, stats, and indexable sitemaps (ARC-956) (ARC-956)
 
+- add public player profiles with game history, stats, and indexable sitemaps (ARC-956) (ARC-956)
 
 ## [1.29.31] - 2026-09-11