+++ CHANGELOG.md
@@ -7,11 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- add Discord community links to desktop navbar and mobile menu (ARC-957)
+
 ## [1.29.32] - 2026-09-11
 
 ### Added
-- add public player profiles with game history, stats, and indexable sitemaps (ARC-956) (ARC-956)
 
+- add public player profiles with game history, stats, and indexable sitemaps (ARC-956) (ARC-956)
 
 ## [1.29.31] - 2026-09-11